Isolate the renderer from the editor code
