+++ src/applications/renderer/package.json
@@ -20,24 +20,13 @@
     "babel-plugin-styled-components": "^1.10.6",
     "lodash": "^4.17.15",
     "prop-types": "15.7.2",
-    "react-select": "^3.0.8",
-    "react-slick": "^0.25.2",
-    "react-slider": "^1.0.2",
-    "redux-devtools-extension": "^2.13.8",
-    "redux-saga": "^1.1.3",
-    "reselect": "^4.0.0",
-    "slick-carousel": "^1.8.1",
     "styled-components": "^4.4.1",
     "vega": "^5.9.0",
-    "vega-tooltip": "0.8.0",
-    "vizzuality-redux-tools": "^4.1.1"
+    "vega-tooltip": "0.8.0"
   },
   "peerDependencies": {
     "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "react-redux": "^7.1.3",
-    "redux": "^4.0.4",
-    "redux-thunk": "^2.3.0"
+    "react-dom": "^16.12.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.7.5",
