--- src/applications/renderer/src/components/chart-title/index.js
@@ -1,3 +1,0 @@
-import ChartTitle from './component';
-
-export default ChartTitle;
