--- src/applications/renderer/src/components/chart-title/style.js
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const StyledTitle = styled.h3`
-  flex-shrink: 0;
-  padding: 5px 0 0 0;
-  color: #393F44;
-  font-size: 16px;
-  text-align: center;
-`;
