+++ src/applications/renderer/src/components/chart/component.js
@@ -1,17 +1,13 @@
-import React, { Fragment, Suspense } from "react";
+import React from "react";
 import PropTypes from 'prop-types';
 import * as vega from "vega";
 import { vega as vegaTooltip }  from "vega-tooltip";
 
-import { JSTypes } from "@widget-editor/types";
-
 import isEqual from "lodash/isEqual";
 import debounce from "lodash/debounce";
-import has from "lodash/has";
 
-import { StyledContainer, ChartNeedsOptions } from "./styles";
 
-const ColumnSelections = React.lazy(() => import("../column-selections"));
+import { StyledContainer } from "./styles";
 
 const getTooltipConfigFields = (widget) => {
   const vegaConfig = widget;
@@ -47,10 +43,6 @@ class Chart extends React.Component {
   constructor(props) {
     super(props);
     this.vega = null;
-    this.state = {
-      chartReady: false
-    }
-    this.standalone = props.standalone || false;
     this.handleResize = debounce(this.handleResize.bind(this), 250);
   }
 
@@ -60,11 +52,9 @@ class Chart extends React.Component {
   }
 
   componentDidUpdate(prevProps) {
-    const widgetChanged = !isEqual(prevProps.widget, this.props.widget);
-    const standaloneConfigChanged =
-      !isEqual(prevProps.standaloneConfiguration, this.props.standaloneConfiguration);
+    const widgetConfigChanged = !isEqual(prevProps.widgetConfig, this.props.widgetConfig);
 
-    if (widgetChanged || standaloneConfigChanged) {
+    if (widgetConfigChanged) {
       this.generateVegaChart();
     }
   }
@@ -73,7 +63,6 @@ class Chart extends React.Component {
     if (this.vega) {
       window.removeEventListener("resize", this.handleResize);
       this.vega = null;
-      this.setState({ chartReady: false, invalidData: false });
       this.chart.innerHTML = '';
     }
   }
@@ -97,13 +86,17 @@ class Chart extends React.Component {
   handleResize() {
     const { view } = this;
     if (view) {
-      this.setSize();
-      if (this.vega) {
-      this.vega
-        .width(this.width)
-        // .height(this.height) // This is a test, currently the renderer resizes its height
-        .run();
-      }
+      // This requestAnimationFrame makes sure that the size is computed when the browser has
+      // finished computing the layout
+      requestAnimationFrame(() => {
+        this.setSize();
+
+        if (this.vega) {
+          this.vega
+            .width(this.width)
+            .run();
+        }
+      });
     }
   }
 
@@ -122,7 +115,7 @@ class Chart extends React.Component {
 
   generateRuntime(configuration) {
     const { chart } = this;
-    this.setSize();
+
     if (chart) {
       try {
         // To avoid memory leaks, the view is destroyed when a new one is created
@@ -180,151 +173,52 @@ class Chart extends React.Component {
     }
   }
 
-  noDataAvailable() {
-    return !this.standalone &&
-      !this.props.advanced &&
-      (!this.props.editor.widgetData || this.props.editor.widgetData.length === 0);
-  }
-
-  // XXX: makes sure custom charts has nessesary info to render
-  verifyCustomChart(conf) {
-    const axisY = conf?.config?.axisY || {};
-    delete axisY.labelAlign;
-    delete axisY.labelBaseline;
-    const config = has(conf, "config") ? {
-      ...conf.config,
-        axisY: {
-          ...axisY,
-          minExtent: 40
-        }
-    } : {}
-    return {
-      ...conf,
-      config
-    }
-  }
-
   generateVegaChart() {
-    const {
-      advanced,
-      thumbnail,
-      widget: vegaConfiguration,
-      standaloneConfiguration,
-    } = this.props;
+    const { thumbnail, widgetConfig } = this.props;
 
-    if (this.noDataAvailable()) {
-      if (this.vega) {
-        window.removeEventListener("resize", this.handleResize);
-        this.vega = null;
-        this.setState({ chartReady: false, invalidData: true });
-      }
-      return;
-    } else {
-      this.setState({ invalidData: false });
-    }
+    // This requestAnimationFrame makes sure that the size is computed when the browser has finished
+    // computing the layout
+    requestAnimationFrame(() => {
+      this.setSize();
 
-    if (this.standalone && standaloneConfiguration) {
       if (thumbnail) {
-        let clearAxis = standaloneConfiguration;
-        delete clearAxis.axisX;
-        delete clearAxis.axisY;
-        delete clearAxis.axes;
-        delete clearAxis.axis;
-        this.generateRuntime(clearAxis);
+        const widgetConfigWithoutAxes = { ...widgetConfig };
+        delete widgetConfigWithoutAxes.axisX;
+        delete widgetConfigWithoutAxes.axisY;
+        delete widgetConfigWithoutAxes.axes;
+        delete widgetConfigWithoutAxes.axis;
+        this.generateRuntime(widgetConfigWithoutAxes);
       } else {
-        this.generateRuntime(this.verifyCustomChart(standaloneConfiguration));
+        this.generateRuntime(widgetConfig);
       }
-    } else {
-      const editedConfiguration = { ...vegaConfiguration };
-
-      // XXX: Remove any nessesary information if not advanced mode
-      // This is for example if:
-      // 1. User deletes the entire custom configuration
-      if (!advanced) {
-        delete editedConfiguration.legends;
-      }
-
-      // If not standalone, we don't want to show the axes titles
-      editedConfiguration.axes = [...editedConfiguration.axes];
-      editedConfiguration.axes?.forEach((axis, index) => {
-        if (axis.title) {
-          editedConfiguration.axes[index] = { ...axis };
-          delete editedConfiguration.axes[index].title;
-        }
-      });
-
-      this.generateRuntime(editedConfiguration);
-    }
-  }
-
-  columnSelection() {
-    return (
-      <Suspense fallback={<div>Loading...</div>}>
-        <ColumnSelections compact={this.props.compact} />
-      </Suspense>
-    );
+    });
   }
 
   render() {
-    const { thumbnail, standalone, advanced = false, configuration: { chartType } } = this.props;
-    const { chartReady, invalidData } = this.state;
+    const { thumbnail } = this.props;
 
     return (
       <StyledContainer
-        standalone={standalone}
+        className="c-renderer"
         thumbnail={thumbnail}
-        compact={this.props.compact}
-        hasYAxis={chartType !== "pie" && chartType !== "donut"}
         ref={(c) => {
           this.view = c;
         }}
       >
-        {!this.noDataAvailable() && (
-          <div
-            className="c-chart"
-            ref={(c) => {
-              this.chart = c;
-            }}
-          />
-        )}
-
-        {this.noDataAvailable() && (
-          <Fragment>
-            {chartReady && (
-              <ChartNeedsOptions>
-                Select value & category to visualize data
-              </ChartNeedsOptions>
-            )}
-            {invalidData && (
-              <ChartNeedsOptions>
-                No data available
-              </ChartNeedsOptions>
-            )}
-          </Fragment>
-        )}
-
-        {!this.standalone && !advanced && this.columnSelection()}
+        <div
+          className="c-chart"
+          ref={(c) => {
+            this.chart = c;
+          }}
+        />
       </StyledContainer>
     );
   }
 }
 
 Chart.propTypes = {
-  editor: PropTypes.shape({
-    widgetData: PropTypes.array
-  }),
-  configuration: JSTypes.configuration,
-  standalone: PropTypes.bool,
   thumbnail: PropTypes.bool,
-  compact: PropTypes.any,
-  widget: PropTypes.object,
-  standaloneConfiguration: PropTypes.object,
-  advanced: PropTypes.bool
-};
-
-Chart.defaultProps = {
-  width: 0,
-  height: 0,
+  widgetConfig: PropTypes.object,
 };
 
 export default Chart;
