+++ src/applications/renderer/src/components/legend/legend-close.js
