--- src/applications/renderer/src/components/standalone/index.js
@@ -1,3 +1,0 @@
-import StandaloneComponent from "./component";
-
-export default StandaloneComponent;
