+++ src/applications/widget-editor/src/components/column-selections/component.js
