+++ src/applications/widget-editor/src/components/column-selections/index.js
