+++ src/applications/widget-editor/src/components/column-selections/style.js
