+++ src/applications/widget-editor/src/components/column-selections/utils.js
