+++ src/applications/widget-editor/src/components/legend/style.js
