+++ src/applications/widget-editor/src/components/legend/utils.js
