+++ src/applications/widget-editor/src/components/select-chart/component.js
