+++ src/applications/widget-editor/src/components/select-chart/components/ChartIcon/style.js
