+++ src/applications/widget-editor/src/components/select-chart/components/ChartMenu/component.js
