+++ src/applications/widget-editor/src/components/select-chart/const.js
