+++ src/applications/widget-editor/src/components/select-chart/index.js
