+++ src/applications/widget-editor/src/components/select-chart/style.js
